firstdraft
I'm not really sure how github works but I hope I did this right(I couldn't see the html draft so I just made mine).
Basically this is it so check it out and see if there is any changes that needs to be made. The background doesn't look the same as in the design, I have no idea why.